SEDONA-743 Add geom from mysql function.
